+++ src/errors/mod.rs
@@ -3,7 +3,6 @@ mod collection_collision;
 mod collection_error;
 mod item_collision;
 mod result_collection_error;
-mod tuple_collection_error;
 mod tuple_extension_error;
 mod unzip_error;
 
@@ -12,6 +11,5 @@ pub use collection_collision::*;
 pub use collection_error::*;
 pub use item_collision::*;
 pub use result_collection_error::*;
-pub use tuple_collection_error::*;
 pub use tuple_extension_error::*;
 pub use unzip_error::*;