+++ tests/errors/main.rs
@@ -4,6 +4,5 @@ mod capacity_mismatch;
 mod collection_collision;
 mod collection_error;
 mod result_collection_error;
-mod tuple_collection_error;
 mod tuple_extension_error;
 mod unzip_error;